@@ -2,7 +2,7 @@ name: scrumpoker
 description: Scrum Poker
 version: 1.0.0+1
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
 dependencies:
   flutter:
     sdk: flutter
